+++ cpp/include/cudf/io/writers.hpp
@@ -87,9 +87,10 @@ class writer {
    * @brief Writes the entire dataset.
    *
    * @param table Set of columns to output
+   * @param metadata Table metadata and column names
    * @param stream Optional stream to use for device memory alloc and kernels
    */
-  void write_all(table_view const& table, cudaStream_t stream = 0);
+  void write_all(table_view const& table, const table_metadata *metadata = nullptr, cudaStream_t stream = 0);
 };
 
 }  // namespace orc
@@ -148,9 +149,10 @@ class writer {
    * @brief Writes the entire dataset.
    *
    * @param table Set of columns to output
+   * @param metadata Table metadata and column names
    * @param stream Optional stream to use for device memory alloc and kernels
    */
-  void write_all(table_view const& table, cudaStream_t stream = 0);
+  void write_all(table_view const& table, const table_metadata *metadata = nullptr, cudaStream_t stream = 0);
 };
 
 }  // namespace parquet